requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
pdfminer.six
PyPDF2
regex
python-dateutil
click
